backend/requirements.txt

flask==2.2.3
flask-cors==3.0.10
python-dotenv==1.0.0
supabase==1.0.3
pyjwt==2.6.0
bcrypt==4.0.1
gunicorn==20.1.0
pytest==7.3.1
pytest-flask==1.2.0
